--- src/main/java/com/bipa/bizsurvey/infra/s3/ShareType.java
@@ -1,15 +1,0 @@
-package com.bipa.bizsurvey.infra.s3;
-
-public enum ShareType {
-    INTERNAL("internal"),
-    EXTERNAL("external");
-    private final String type;
-
-    ShareType(String type) {
-        this.type = type;
-    }
-
-    public String getType() {
-        return type + "-";
-    }
-}
